move new warning to StartPage
